server/requirements.txt
fastapi
pydantic
uvicorn
requests
supabase
numpy
passlib
python-dotenv
passlib
websockets
